Finalizing Comments Part 2 of 2!
Functional comments are added.

Also added additional raise calls whenever text files are read from to see if they even exist.

When a text file needs to be written to, programs will now check to see if the file exists first. If it does it will do a standard write that overwrites the previous contents, otherwise it will create the file then write to it.

Fixed small indentation error in arduino sketch